infra: add docker
