+++ app/[[...slug]]/client.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import React from 'react'
 import Link from 'next/link'
 
 export function ClientOnly() {
