+++ app/[[...slug]]/page.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { ClientOnly } from './client'
 
 export function generateStaticParams() {
