+++ app/editor/page.jsx
@@ -1,42 +1,15 @@
-'use client'
-import { useRef } from 'react';
-import { Editor as MonacoEditor } from '@monaco-editor/react';
-import * as Y from 'yjs';
-import { WebsocketProvider } from 'y-websocket';
-import { MonacoBinding } from 'y-monaco';
+import dynamic from 'next/dynamic';
 
-function Editor() {
-  const editorRef = useRef(null);
+const Editor = dynamic(() => import('./Editor'), {
+  ssr: false,
+});
 
-  function editorOnMount(editor) {
-    editorRef.current = editor;
-
-    const doc = new Y.Doc();
-    // const provider = new WebsocketProvider('wss://demos.yjs.dev', 'monaco-demo', doc);
-    const provider = new WebsocketProvider(
-      'ws://127.0.0.1:13701',
-      'monaco-demo',
-      doc
-    );
-    const type = doc.getText('monaco');
-
-    new MonacoBinding(
-      type,
-      editorRef.current.getModel(),
-      new Set([editorRef.current]),
-      provider.awareness
-    );
-  }
-
+function Page() {
   return (
-    <MonacoEditor
-      height="100vh"
-      width="100vw"
-      theme="vs-dark"
-      defaultLanguage="cpp"
-      onMount={editorOnMount}
-    />
+    <div>
+      <Editor />
+    </div>
   );
 }
 
-export default Editor;
+export default Page;
