+++ app/main/menuIcon.tsx
@@ -1,5 +1,3 @@
-import React from "react"
-
 type Props = {
     item: string
 }
