+++ app/main/page.tsx
@@ -1,5 +1,4 @@
 'use client'
-import React from "react";
 import Navbar from "./navbar";
 import SideMenu from "./sideMenu";
