+++ next.config.mjs
@@ -1,8 +1,7 @@
-/** @type {import('next').NextConfig} */
 const nextConfig = {
-    output: 'export', // Outputs a Single-Page Application (SPA).
-    distDir: './dist', // Changes the build output directory to `./dist/`.
-    basePath: process.env.NEXT_PUBLIC_BASE_PATH,
-  }
+  output: 'standalone', // Outputs a Single-Page Application (SPA).
+  distDir: './dist', // Changes the build output directory to `./dist/`.
+  basePath: process.env.NEXT_PUBLIC_BASE_PATH,
+};
 
-  export default nextConfig
+export default nextConfig;
